README.md: remove mention of Apache v2.0 license
In commit 09d45c95de9a6cdd8949892129cb76161b958d3e, the
SOURCE_CODE_LICENSE file is deleted.

We assume that the goal is to move the whole project under SIL instead
of Apache v2.0. This is why the mention of the Apache v2.0 license for
source code should be removed, as the Apache v2.0 license file has been
removed.