Added new powerline symbols
I needed new symbols to make my vim-airline look like a bevel, so I added them. I added symbols from E0B8 to E0BF. I also noticed that E0B1 and E0B3 have a corner that is a bit off, and the tip is also miscentered making the glyph asymmetrical. Not a big deal but I wanted to mention it